# server/solar_server/app.py
"""HTTP routes. SOLAR_ROLE picks which half of the app a service exposes."""

from __future__ import annotations

import logging
import os
import time
from collections.abc import Callable
from datetime import UTC, datetime
from pathlib import Path
from typing import Annotated, Any, Literal

from fastapi import FastAPI, Header, HTTPException, Query, Request, Response
from fastapi.responses import JSONResponse
from fastapi.staticfiles import StaticFiles

from solar_server import health, tasks, views
from solar_server.auth import TokenVerifier, check_plug_key, check_scheduler, google_verifier
from solar_server.config import Settings
from solar_server.schema import Accuracy, Co2, Daily, Now, PlugReport, Timeline
from solar_server.sources import LiveSources, Sources
from solar_server.store import FirestoreStore, Store

WEB_DIST = Path(
    os.environ.get("SOLAR_WEB_DIST", Path(__file__).resolve().parents[2] / "web" / "dist")
)


def create_app(
    settings: Settings,
    store: Store,
    sources: Sources,
    verify: TokenVerifier = google_verifier,
    clock: Callable[[], float] = time.time,
) -> FastAPI:
    app = FastAPI(
        title="Solar Farm",
        docs_url=None,
        redoc_url=None,
        # One schema per model, so the generated TypeScript names match.
        separate_input_output_schemas=False,
    )

    @app.get("/health")
    def health() -> dict:
        return {"ok": True, "role": settings.role}

    if settings.role == "edge":
        _edge_routes(app, settings, store, sources, verify, clock)
    else:
        _web_routes(app, store, clock)
    return app


# Cloud Scheduler gives a collect up to 120 s (infra/scheduler.tf).
COLLECT_LEASE = 120


def _edge_routes(app, settings, store, sources, verify, clock) -> None:
    def now() -> datetime:
        return datetime.fromtimestamp(clock(), UTC)

    @app.get("/plug/plan")
    def plug_plan(x_plug_key: Annotated[str | None, Header()] = None) -> dict:
        check_plug_key(x_plug_key, settings.plug_key)
        p = tasks.plug_plan(store)
        if p is None:
            raise HTTPException(status_code=503, detail="no plan yet")
        return p

    @app.post("/plug/report", status_code=204, response_model=None)
    def plug_report(
        report: PlugReport,
        x_plug_key: Annotated[str | None, Header()] = None,
        plan: bool = False,
    ) -> Response | dict:
        """Store a report; ``?plan=1`` also returns the current plan when available."""
        check_plug_key(x_plug_key, settings.plug_key)
        item: dict[str, Any] = report.model_dump(exclude_none=True)
        item["received"] = int(clock())
        tasks.record_plug(store, item)
        p = tasks.plug_plan(store) if plan else None
        return Response(status_code=204) if p is None else JSONResponse(p)

    @app.post("/tasks/collect")
    def collect(request: Request) -> dict:
        check_scheduler(request, settings.scheduler_sa, verify)
        t = now()
        log = logging.getLogger(__name__)
        # Prevent overlapping scheduled collects while the lease is held.
        if not store.claim("collect", int(t.timestamp()), COLLECT_LEASE):
            log.warning("previous collect still running; skipped")
            return {"t": int(t.timestamp()), "skipped": True}
        try:
            sample = tasks.collect(store, sources, t)
            try:
                tasks.plan(store, sources, settings, t, archive=tasks.archive_due(t))
            except Exception:
                log.exception("forecast refresh failed; replanning with the cached one")
                try:
                    tasks.refresh_charging_plan(store, settings, t)
                except Exception:
                    log.exception("charging replan failed after telemetry collection")
            health.check(store, int(clock()))
        finally:
            store.release("collect")
        return sample

    @app.post("/tasks/plan")
    def plan(request: Request) -> dict:
        check_scheduler(request, settings.scheduler_sa, verify)
        p = tasks.plan(store, sources, settings, now())
        return {"windows": p["windows"], "generated_at": p["generated_at"]}


def _web_routes(app, store, clock) -> None:
    # Preserve omitted keys instead of filling them with model defaults.
    def route(path: str, model: type):
        return app.get(path, response_model=model, response_model_exclude_unset=True)

    @route("/api/now", Now)
    def api_now() -> dict:
        return views.now_view(store, int(clock()))

    @route("/api/timeline", Timeline)
    def api_timeline(past_hours: Annotated[int, Query(ge=1, le=168)] = 24) -> dict:
        return views.timeline_view(store, int(clock()), past_hours)

    @route("/api/accuracy", Accuracy)
    def api_accuracy(
        lead_hours: Annotated[int, Query(ge=1, le=23)] = 6,
        past_hours: Annotated[int, Query(ge=1, le=168)] = 24,
    ) -> dict:
        return views.accuracy_view(store, int(clock()), lead_hours, past_hours)

    @route("/api/daily", Daily)
    def api_daily(days: Annotated[int, Query(ge=1, le=60)] = 14) -> dict:
        return views.daily_view(store, int(clock()), days)

    @route("/api/co2", Co2)
    def api_co2(
        by: Literal["day", "week", "month"] = "day",
        count: Annotated[int, Query(ge=1, le=60)] = 14,
    ) -> dict:
        return views.co2_view(store, int(clock()), by, count)

    if WEB_DIST.is_dir():
        app.mount("/", StaticFiles(directory=WEB_DIST, html=True), name="web")
    else:

        @app.get("/", include_in_schema=False)
        def placeholder() -> dict:
            return {"message": "Dashboard not built yet; the API is at /api/now."}


def main() -> FastAPI:
    """Entry point for uvicorn --factory."""
    logging.basicConfig(level=logging.INFO)
    settings = Settings.from_env()
    return create_app(settings, FirestoreStore(settings.project), LiveSources(settings))